Fixes issue 2169. Thanks goes to Dominik for the point and the fix.
git-svn-id: http://csipsimple.googlecode.com/svn/trunk/CSipSimple@2076 9f815046-5998-e9c0-b7e2-ac03a23edfa4

+++ src/com/csipsimple/service/DeviceStateReceiver.java
@@ -84,7 +84,7 @@ public class DeviceStateReceiver extends BroadcastReceiver {
             if (accId == SipProfile.INVALID_ID) {
                 // allow remote side to send us integers.
                 // previous call will warn, but that's fine, no worries
-                accId = intent.getLongExtra(SipProfile.FIELD_ID, SipProfile.INVALID_ID);
+                accId = intent.getIntExtra(SipProfile.FIELD_ID, (int) SipProfile.INVALID_ID);
             }
 
             if (accId != SipProfile.INVALID_ID) {
